Move monthly daily EV energy usage to transportation calculations file and reducer
